match-details: breadcrumb quartet, drop footer + public route, rename page
- Extract the header breadcrumb into a colocated quartet
  (breadcrumb.tsx/.page/.factory/.test)
- Remove the page footer/logo entirely (+ dead .md-footer/.md-logo CSS)
- Rename the match-details-page.tsx husk to match-details.tsx (`MatchDetails`)
  and merge its legacy suite into match-details.test.tsx
- Delete the public /p/matches/$matchId route and tear out the `standalone`
  rendering mode from MatchDetails / Breadcrumb / MatchDetailsError
- Preload match details on row hover/intent (router defaultPreload + route
  loaders priming the React Query cache)
- Drop stale /p/matches references from the get_match docstring, a test
  docstring, and the generated schema.d.ts
