+++ web-client/src/components/matches/match-details/breadcrumb.factory.tsx
@@ -1,0 +1,11 @@
+import type { BreadcrumbProps } from './breadcrumb'
+
+/** Props for `Breadcrumb`. */
+export function buildBreadcrumbProps(
+  overrides: Partial<BreadcrumbProps> = {},
+): BreadcrumbProps {
+  return {
+    matchId: 'abcdef0000',
+    ...overrides,
+  }
+}
